# docs/swagger.json
{
  "components": {
    "schemas": {
      "data": {
        "title": "data Schema",
        "properties": { "data": { "$ref": "#/components/schemas/Wallet" } },
          "title": "properties Schema",
        "type": "object"
      },
      "ErrorResponse": {
        "title": "ErrorResponse Schema",
        "description": "Application error response",
        "properties": {
          "code": {
            "enum": [400, 401, 404, 500, 502, 422, 403, 409, 429, 501],
            "example": 400,
            "type": "integer"
          },
          "error": { "example": "field: required", "type": "string" },
          "message": {
            "enum": [
              "Invalid request parameters",
              "Authorization failed",
              "Resource not found",
              "Internal server error",
              "Database error occurred",
              "External service error",
              "Error rendering response",
              "Access forbidden",
              "Resource conflict",
              "Too many requests",
              "Unsupported operation"
            ],
            "example": "Invalid request parameters",
            "type": "string"
          },
          "type": { "$ref": "#/components/schemas/ErrorType" }
        },
        "type": "object"
      },
      "ErrorType": {
        "title": "ErrorType Schema",
        "type": "string",
        "x-enum-varnames": [
          "ErrorTypeValidation",
          "ErrorTypeDatabase",
          "ErrorTypeAuthorization",
          "ErrorTypeNotFound",
          "ErrorTypeInternal",
          "ErrorTypeExternalService",
          "ErrorTypeRender",
          "ErrorTypeForbidden",
          "ErrorTypeConflict",
          "ErrorTypeRateLimit",
          "ErrorTypeUnsupported"
        ]
      },
      "Response": {
        "title": "Response Schema",
        "allOf": [{ "$ref": "#/components/schemas/data" }],
        "description": "Standard API response wrapper",
        "properties": {
          "data": {},
          "message": {
            "enum": [
              "Success",
              "Resource created successfully",
              "Resource updated successfully",
              "Resource deleted successfully"
            ],
            "example": "Success",
            "type": "string"
          },
          "meta": {
            "properties": {
              "count": { "type": "integer" },
              "limit": { "type": "integer" },
              "next_token": { "type": "string" },
              "query": { "type": "string" }
            },
            "type": "object"
          },
          "status": {
            "enum": [200, 202, 204],
            "example": 200,
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "ProjectCreatePayload": {
        "title": "ProjectCreatePayload Schema",
        "description": "Payload for creating a new project",
        "properties": {
          "addressLine1": {
            "example": "123 Main St",
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "addressLine2": {
            "example": "Suite 100",
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "budget": {
            "example": 10000.5,
            "minimum": 0,
            "type": "number",
            "nullable": true
          },
          "city": {
            "example": "New York",
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "country": {
            "example": "US",
            "format": "iso-3166-1-alpha-2",
            "pattern": "^[A-Z]{2}$",
            "type": "string",
            "nullable": true
          },
          "description": {
            "example": "Detailed project description",
            "maxLength": 1000,
            "type": "string",
            "nullable": true
          },
          "endDate": {
            "example": "2024-12-31T00:00:00Z",
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "name": {
            "example": "My Project",
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "startDate": {
            "example": "2024-01-01T00:00:00Z",
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "stateProvince": {
            "example": "NY",
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "status": {
            "default": "ongoing",
            "enum": ["ongoing", "completed", "canceled"],
            "example": "ongoing",
            "type": "string"
          },
          "tags": {
            "example": [
              "123e4567-e89b-12d3-a456-426614174000",
              "123e4567-e89b-12d3-a456-426614174001"
            ],
            "items": { "format": "uuid", "type": "string" },
            "maxItems": 10,
            "type": "array",
            "uniqueItems": true
          },
          "website": {
            "example": "https://example.com",
            "format": "uri",
            "type": "string",
            "nullable": true
          },
          "zipPostalCode": {
            "example": "10001",
            "format": "zip-code",
            "pattern": "^\\d{5}(?:[-\\s]\\d{4})?$",
            "type": "string",
            "nullable": true
          }
        },
        "required": ["name", "status"],
        "type": "object"
      },
      "ProjectUpdatePayload": {
        "title": "ProjectUpdatePayload Schema",
        "description": "Payload for updating an existing project",
        "properties": {
          "addressLine1": {
            "example": "123 Main St",
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "addressLine2": {
            "example": "Suite 100",
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "budget": {
            "example": 10000.5,
            "minimum": 0,
            "type": "number",
            "nullable": true
          },
          "city": {
            "example": "New York",
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "country": {
            "example": "US",
            "format": "iso-3166-1-alpha-2",
            "pattern": "^[A-Z]{2}$",
            "type": "string",
            "nullable": true
          },
          "description": {
            "example": "Detailed project description",
            "maxLength": 1000,
            "type": "string",
            "nullable": true
          },
          "endDate": {
            "example": "2024-12-31T00:00:00Z",
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "name": {
            "example": "My Project",
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "startDate": {
            "example": "2024-01-01T00:00:00Z",
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "stateProvince": {
            "example": "NY",
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "status": {
            "enum": ["ongoing", "completed", "canceled"],
            "example": "ongoing",
            "type": "string"
          },
          "tags": {
            "example": [
              "123e4567-e89b-12d3-a456-426614174000",
              "123e4567-e89b-12d3-a456-426614174001"
            ],
            "items": { "format": "uuid", "type": "string" },
            "maxItems": 10,
            "type": "array",
            "uniqueItems": true
          },
          "website": {
            "example": "https://example.com",
            "format": "uri",
            "type": "string",
            "nullable": true
          },
          "zipPostalCode": {
            "example": "10001",
            "format": "zip-code",
            "pattern": "^\\d{5}(?:[-\\s]\\d{4})?$",
            "type": "string",
            "nullable": true
          }
        },
        "type": "object"
      },
...
      "Wallet": {
        "title": "Wallet Schema",
        "description": "A wallet entity",
        "properties": {
          "balance": { "example": 100.5, "type": "number" },
          "createdAt": { "example": "2023-01-01T00:00:00Z", "type": "string" },
          "currency": { "example": "USD", "type": "string" },
          "name": { "example": "My Wallet", "type": "string" },
          "projectId": {
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "type": "string"
          },
          "tags": {
            "items": { "type": "string" },
            "type": "array",
            "uniqueItems": false
          },
          "updatedAt": { "example": "2023-01-01T00:00:00Z", "type": "string" },
          "userId": {
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "type": "string"
          },
          "walletId": {
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "Bearer token authentication",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": { "email": "support@example.com", "name": "API Support" },
    "description": "REST API for expense tracking application with user management",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService": "http://swagger.io/terms/",
    "title": "Expense Tracker API",
    "version": "1.0"
  },
...
  "paths": {
...
    "/contacts/search": {
      "get": {
        "description": "Searches for Contacts based on a query string",
        "operationId": "SearchContacts",
        "parameters": [
          {
            "description": "Search query",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": { "maxLength": 100, "minLength": 1, "type": "string" }
          },
          {
            "description": "Maximum number of results",
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/data" }],
                  "description": "Standard API response wrapper",
                  "properties": {
                    "data": {},
                    "message": {
                      "enum": [
                        "Success",
                        "Resource created successfully",
                        "Resource updated successfully",
                        "Resource deleted successfully"
                      ],
                      "example": "Success",
                      "type": "string"
                    },
                    "meta": {
                      "properties": {
                        "count": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "next_token": { "type": "string" },
                        "query": { "type": "string" }
                      },
                      "type": "object"
                    },
                    "status": {
                      "enum": [200, 202, 204],
                      "example": 200,
                      "type": "integer"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Unauthorized"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [{ "BearerAuth": [] }],
        "summary": "Search Contacts",
        "tags": ["Contacts"]
      }
    },
...
    "/projects": {
      "get": {
        "description": "Returns a paginated list of projects",
        "operationId": "ListProjectsPaginated",
        "parameters": [
          {
            "description": "Number of projects to return",
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 10,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Token for the next page",
            "in": "query",
            "name": "next_token",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/data" }],
                  "description": "Standard API response wrapper",
                  "properties": {
                    "data": {},
                    "message": {
                      "enum": [
                        "Success",
                        "Resource created successfully",
                        "Resource updated successfully",
                        "Resource deleted successfully"
                      ],
                      "example": "Success",
                      "type": "string"
                    },
                    "meta": {
                      "properties": {
                        "count": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "next_token": { "type": "string" },
                        "query": { "type": "string" }
                      },
                      "type": "object"
                    },
                    "status": {
                      "enum": [200, 202, 204],
                      "example": 200,
                      "type": "integer"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Unauthorized"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [{ "BearerAuth": [] }],
        "summary": "List projects with pagination",
        "tags": ["Projects"]
      },
      "post": {
        "description": "Creates a new project for the authenticated user",
        "operationId": "CreateProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProjectCreatePayload" }
            }
          },
          "description": "project creation request",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/data" }],
                  "description": "Standard API response wrapper",
                  "properties": {
                    "data": {},
                    "message": {
                      "enum": [
                        "Success",
                        "Resource created successfully",
                        "Resource updated successfully",
                        "Resource deleted successfully"
                      ],
                      "example": "Success",
                      "type": "string"
                    },
                    "meta": {
                      "properties": {
                        "count": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "next_token": { "type": "string" },
                        "query": { "type": "string" }
                      },
                      "type": "object"
                    },
                    "status": {
                      "enum": [200, 202, 204],
                      "example": 200,
                      "type": "integer"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [{ "BearerAuth": [] }],
        "summary": "Create a new project",
        "tags": ["Projects"]
      }
    },
    "/projects/{id}": {
      "delete": {
        "description": "Deletes a project by ID",
        "operationId": "DeleteProject",
        "parameters": [
          {
            "description": "project ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Response" }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [{ "BearerAuth": [] }],
        "summary": "Delete a project",
        "tags": ["Projects"]
      },
      "get": {
        "description": "Retrieves a project by ID",
        "operationId": "GetProject",
        "parameters": [
          {
            "description": "project ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/data" }],
                  "description": "Standard API response wrapper",
                  "properties": {
                    "data": {},
                    "message": {
                      "enum": [
                        "Success",
                        "Resource created successfully",
                        "Resource updated successfully",
                        "Resource deleted successfully"
                      ],
                      "example": "Success",
                      "type": "string"
                    },
                    "meta": {
                      "properties": {
                        "count": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "next_token": { "type": "string" },
                        "query": { "type": "string" }
                      },
                      "type": "object"
                    },
                    "status": {
                      "enum": [200, 202, 204],
                      "example": 200,
                      "type": "integer"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [{ "BearerAuth": [] }],
        "summary": "Get a project",
        "tags": ["Projects"]
      },
      "put": {
        "description": "Updates an existing project",
        "operationId": "UpdateProject",
        "parameters": [
          {
            "description": "project ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProjectUpdatePayload" }
            }
          },
          "description": "project update request",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/data" }],
                  "description": "Standard API response wrapper",
                  "properties": {
                    "data": {},
                    "message": {
                      "enum": [
                        "Success",
                        "Resource created successfully",
                        "Resource updated successfully",
                        "Resource deleted successfully"
                      ],
                      "example": "Success",
                      "type": "string"
                    },
                    "meta": {
                      "properties": {
                        "count": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "next_token": { "type": "string" },
                        "query": { "type": "string" }
                      },
                      "type": "object"
                    },
                    "status": {
                      "enum": [200, 202, 204],
                      "example": 200,
                      "type": "integer"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [{ "BearerAuth": [] }],
        "summary": "Update a project",
        "tags": ["Projects"]
      }
    },
...
  },
  "openapi": "3.1.0",
  "servers": [{ "url": "localhost:8080/api/v1" }]
}